% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borrowing.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Borrowed effective sample size}
\usage{
effective_sample_size(object, endpoint = NULL)
}
\arguments{
\item{object}{A fitted \code{safeab_fit} object.}

\item{endpoint}{Optional endpoint name.}
}
\value{
A named numeric vector of borrowed effective sample sizes.
}
\description{
Borrowed effective sample size
}
