% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.safeab_fit}
\alias{plot.safeab_fit}
\title{Plot posterior SAFE-AB dose-response curves}
\usage{
\method{plot}{safeab_fit}(x, endpoint = x$endpoints, n_grid = 101L, ...)
}
\arguments{
\item{x}{A fitted \code{safeab_fit} object.}

\item{endpoint}{Endpoints to plot.}

\item{n_grid}{Number of standardized dose-grid points.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
The fitted object, invisibly.
}
\description{
Plot posterior SAFE-AB dose-response curves
}
