% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{safeab_prior}
\alias{safeab_prior}
\title{Specify priors for SAFE-AB endpoint models}
\usage{
safeab_prior(
  toxicity_mean = c(stats::qlogis(0.1), 0),
  toxicity_sd = c(1.5, 0.75),
  efficacy_mean = c(stats::qlogis(0.15), 0, 0),
  efficacy_sd = c(1.5, 2, 2)
)
}
\arguments{
\item{toxicity_mean, toxicity_sd}{Numeric vectors of length two for the
monotone toxicity model.}

\item{efficacy_mean, efficacy_sd}{Numeric vectors of length three for the
quadratic efficacy model.}
}
\value{
A \code{safeab_prior} object.
}
\description{
Specify priors for SAFE-AB endpoint models
}
