% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.safeab_fit}
\alias{summary.safeab_fit}
\title{Summarize a SAFE-AB fit}
\usage{
\method{summary}{safeab_fit}(object, ...)
}
\arguments{
\item{object}{A fitted \code{safeab_fit} object.}

\item{...}{Unused.}
}
\value{
A \code{summary.safeab_fit} object.
}
\description{
Summarize a SAFE-AB fit
}
